refactor(auth/service): merge core auth service and jwt auth service
