+++ packages/auth/src/core/index.ts
@@ -1,2 +1,2 @@
+export * from '../services';
 export * from './components';
-export * from './services';
