+++ packages/auth/src/index.ts
@@ -1,7 +1,3 @@
 export * from './core';
-// Please refer to REAMDME for more information
-/* JWT export */
-export { AuthService, AuthState, refreshToken, resolveJWTAuthState } from './jwt';
+export * from './services';
 export * from './__mocks__';
-
-/* ***********/
