--- packages/auth/src/jwt/index.ts
@@ -1,1 +1,0 @@
-export * from './services';
