--- packages/auth/src/jwt/services/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthService';
