+++ packages/auth/src/services/AuthService.ts
@@ -23,6 +23,21 @@ export const AuthService = {
       throw new Error(`Login Error: ${error.message}`);
     }
   },
+
+  loginWithJWT: async (token: string) => {
+    try {
+      Http.setToken(token);
+
+      const response = await Http.post<SigninResponse, any>('/auth/refresh');
+
+      Http.setToken(response.data.access_token);
+
+      return response.data;
+    } catch (error) {
+      throw error;
+    }
+  },
+
   logout: async () => {
     try {
       await Http.post<SignoutResponse>('/auth/logout');
