+++ packages/auth/src/services/interfaces/AuthModel.ts
